% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{auto_config_tf}
\alias{auto_config_tf}
\title{Auto Configurate TensforFlow}
\usage{
auto_config_tf()
}
\description{
Load and install the r-tensorflow conda environment (python / tensorflow can be setup in other ways with reticulate).
}
