% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impulse_fitting.R
\name{estimate_one_timecourse_params_tf}
\alias{estimate_one_timecourse_params_tf}
\title{Estimate One Timecourses Parameters}
\usage{
estimate_one_timecourse_params_tf(
  one_timecourse,
  a_tc_id,
  use_prior,
  n_initializations,
  model,
  prior_pars = NULL,
  fit_intercept = FALSE,
  learning_rate = 0.1,
  n_iterations = 100,
  verbose = FALSE
)
}
\arguments{
\item{one_timecourse}{dynamics of a single timecourse including
time and abundance variables}

\item{a_tc_id}{timecourse id for the timecourse being fit}

\item{use_prior}{If FALSE, fit least squares. If TRUE, add priors for a MAP estimate.}

\item{n_initializations}{Number of initializations to use for each timecourse.}

\item{model}{model to fit:
\describe{
  \item{sigmoid}{one sigmoidal response}
  \item{impulse}{two sigmoidal responses}
  }}

\item{prior_pars}{Named numeric vector of parameters to use for priors (if use_prior is TRUE)
\describe{
  \item{v_sd}{Gaussian SD of assymptotes (v_inter and v_final)}
  \item{rate_shape}{Shape of rate Gamma}
  \item{rate_scale}{Scale of rate Gamma}
  \item{time_shape}{Shape of t_rise and t_fall - t_rise Gamma}
  \item{time_scale}{Scale of t_rise and t_fall - t_rise Gamma}
  }}

\item{fit_intercept}{If TRUE, the intercept will be fit, if FALSE, the intercept will be constrainted to zero}

\item{learning_rate}{learning rate for the Adams optimizer}

\item{n_iterations}{the number of iterations to run the optimizer}

\item{verbose}{if TRUE then print additional information}
}
\description{
Estimate One Timecourses Parameters
}
