% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impulse_fitting.R
\name{estimate_timecourse_params_tf}
\alias{estimate_timecourse_params_tf}
\title{Fit Timecourses}
\usage{
estimate_timecourse_params_tf(measurements, model = "sigmoid",
  n_initializations = 100, use_prior = TRUE, prior_pars = c(v_sd =
  1.2, rate_shape = 2, rate_scale = 0.25, time_shape = 2, time_scale = 25),
  verbose = FALSE)
}
\arguments{
\item{measurements}{a tibble containing:
\itemize{
  \item{\code{tc_id}: a unique indicator for each timecourse"},
  \item{\code{time}: a numeric predictor variable},
  \item{\code{abundance}: a numeric response variable}
  }}

\item{model}{model to fit:
\itemize{
  \item{\code{sigmoid}: one sigmoidal response},
  \item{\code{impulse}: two sigmoidal responses}
  }}
}
\value{
a timecourse list:
\itemize{
  \item{\code{invalid_timecourse_fits}: tibble of parameter initializations for initializations that went to NaN for debugging},
  \item{\code{loss}: tibble of losses for each tc_id and inititalization (init_id)},
  \item{\code{parameters}: tibble of final parameters for each tc_id and initialization (init_id)}
  }
}
\description{
Fit Timecourses
}
\examples{
timecourses <- simulate_timecourses(n = 5)

timecourses \%>\%
  tidyr::unnest(measurements) \%>\%
  # separate by true model
  tidyr::nest(-true_model, .key = "measurements") \%>\%
  # fit all models to each timecourse
  tidyr::crossing(tibble::tibble(model = c("sigmoid", "impulse"))) \%>\%
  dplyr::mutate(timecourse_params = purrr::map2(measurements, model, estimate_timecourse_params_tf, n_initializations = 10))

sigmoids <- fit_timecourses_tensorflow(timecourses, model = "sigmoid", n_initializations = 50, use_prior = FALSE)
impulses <- fit_timecourses_tensorflow(timecourse_subset, model = "impulse", n_initializations = 50, use_prior = FALSE)

sigmoids <- fit_timecourses_tensorflow(timecourse_subset, model = "sigmoid", n_initializations = 60, use_prior = TRUE, verbose = TRUE)
impulses <- fit_timecourses_tensorflow(timecourse_subset, model = "impulse", n_initializations = 60, use_prior = TRUE, verbose = TRUE)

}
