% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_timecourse}
\alias{example_timecourse}
\title{Example Timecourse}
\format{
A tibble with 24 rows and 3 columns:
\describe{
  \item{tc_id}{A number representing a distinct timecourse}
  \item{time}{Time (in divisions)}
  \item{abundance}{Relative expression of the reporter}
  }
}
\usage{
example_timecourse
}
\description{
An example timecourse of fluorescent reporter dynamics in yeast.
}
\keyword{datasets}
