% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impulse_plotting.R
\name{gene_tf_fits}
\alias{gene_tf_fits}
\title{Gene x TF Timecourse}
\usage{
gene_tf_fits(timecourses, label.var = "tc_id")
}
\arguments{
\item{timecourses}{a tibble with each row being model, tc_id, params and timecourses}

\item{label.var}{an additional field to use for labelling timecourses}
}
\value{
a ggplot2 plot
}
\description{
Gene x TF Timecourse
}
