% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_comparison.R
\name{impulse_sigmoid_comparison}
\alias{impulse_sigmoid_comparison}
\title{Impulse Sigmoid Comparison}
\usage{
impulse_sigmoid_comparison(timecourse_losses, fdr_cutoff = 0.001)
}
\description{
Impulse Sigmoid Comparison
}
