% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impulse_plotting.R
\name{kinetics_plotting}
\alias{kinetics_plotting}
\title{Kinetics plotting}
\usage{
kinetics_plotting(augmented_timecourses, saturation = 0.9, max_time,
  fit_timepoints = 100)
}
\description{
Kinetics plotting
}
