% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impulse_fitting.R
\name{reduce_best_timecourse_params}
\alias{reduce_best_timecourse_params}
\title{Reduce to best timecourse parameters}
\usage{
reduce_best_timecourse_params(timecourse_list,
  reduction_type = "loss-min", sufficiency_tolerance = 0.05)
}
\arguments{
\item{timecourse_list}{List output from \code{\link{estimate_timecourse_parameters_tf}}}

\item{reduction_type}{How to choose the best parameter set, options are:
\itemize{
 \item{\code{loss-min}: lowest loss function},
 \item{\code{loss-small-v-small}: loss within \code{sufficiency_tolerance} of minimum loss and then minimize absolute sum of \eqn{v_{inter}} and \eqn{v_{final}} (useful primarily when not using priors).}
}}

\item{sufficiency_tolerance}{All timecourses within 1 + sufficiency_tolerance best fitting parameter set are deemed sufficient}
}
\value{
a list containing top parameter set and losses
}
\description{
Across multiple fits of a timecourse summarize the best fitting timecourse in terms of least-squares error as well as by lowest absolute V within a tolerance of the least-squares set.
}
