% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_parameters}
\alias{simulate_parameters}
\title{Simulate parameters}
\usage{
simulate_parameters(n, model = "sigmoid", prior_pars = c(v_sd = 1.2,
  rate_shape = 2, rate_scale = 0.25, time_shape = 2, time_scale = 25))
}
\arguments{
\item{n}{# of parameter sets to simulate}

\item{model}{name of model to simulate}

\item{prior_pars}{prior parameters for model}
}
\description{
Simulate parameters
}
\examples{
simulate_parameters(n = 20)
}
