% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impulse_fitting.R
\name{summarize_top_timecourses}
\alias{summarize_top_timecourses}
\title{Summarize Top Timecourses}
\usage{
summarize_top_timecourses(combined_timecourses,
  sufficiency_tolerance = 0.05)
}
\arguments{
\item{combined_timecourses}{List output from fit_timecourses_tensorflow}

\item{sufficiency_tolerance}{All timecourses within 1 + sufficiency_tolerance best fitting parameter set are deemed sufficient}
}
\value{
a list containing top timecourses, parameter set and losses
}
\description{
Across multiple fits of a timecourse summarize the best fitting timecourse in terms of least-squares error as well as by lowest absolute V within a tolerance of the least-squares set.
}
