% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_setup.R
\name{timecourse_fit_job_submission}
\alias{timecourse_fit_job_submission}
\title{Setup timecourse fit jobs}
\usage{
timecourse_fit_job_submission(expression_data, njobs = 200, remote_path)
}
\description{
Setup timecourse fit jobs
}
\examples{
timecourse_fit_job_submission(expression_data, njobs = 50, remote_path = "sean@sci-vm-005:/home/sean/data/McIsaac_timecourses/timecourse_data.Rds")

}
