% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impulse_fitting.R
\name{estimate_one_timecourse_params_tf}
\alias{estimate_one_timecourse_params_tf}
\title{Estimate One Timecourses Parameters}
\usage{
estimate_one_timecourse_params_tf(
  one_timecourse,
  a_tc_id,
  use_prior,
  n_initializations,
  model,
  prior_pars = NULL,
  learning_rate = 0.1,
  verbose = FALSE
)
}
\arguments{
\item{one_timecourse}{dynamics of a single timecourse including
time and abundance variables}

\item{a_tc_id}{timecourse id for the timecourse being fit}

\item{use_prior}{If FALSE, fit least squares. If TRUE, add priors for a MAP estimate.}

\item{n_initializations}{Number of initializations to use for each timecourse.}

\item{model}{model to fit:
\itemize{
  \item{\code{sigmoid}: one sigmoidal response},
  \item{\code{impulse}: two sigmoidal responses}
  }}

\item{prior_pars}{Named numeric vector of parameters to use for priors (if use_prior is TRUE)
\itemize{
  \item{\code{v_sd}: Gaussian SD of assymptotes (v_inter and v_final)}
  \item{\code{rate_shape}: Shape of rate Gamma}
  \item{\code{rate_scale}: Scale of rate Gamma}
  \item{\code{time_shape}: Shape of t_rise and t_fall - t_rise Gamma}
  \item{\code{time_scale}: Scale of t_rise and t_fall - t_rise Gamma}
  }}

\item{learning_rate}{learning rate for the Adams optimizer}

\item{verbose}{if TRUE then print additional information}
}
\description{
Estimate One Timecourses Parameters
}
