% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impulse_fitting.R
\name{estimate_timecourse_params_tf}
\alias{estimate_timecourse_params_tf}
\title{Fit Timecourses}
\usage{
estimate_timecourse_params_tf(measurements, model = "sigmoid",
  n_initializations = 100, use_prior = TRUE, prior_pars = c(v_sd =
  1.2, rate_shape = 2, rate_scale = 0.25, time_shape = 2, time_scale = 25),
  verbose = FALSE)
}
\arguments{
\item{measurements}{a tibble containing:
\itemize{
  \item{tc_id: a unique indicator for each timecourse"},
  \item{time: a numeric predictor variable},
  \item{abundance: a numeric response variable}
  }}

\item{model}{model to fit:
\itemize{
  \item{sigmoid: one sigmoidal response},
  \item{impulse: two sigmoidal responses}
  }}
}
\description{
Fit Timecourses
}
\examples{
timecourses <- simulate_timecourses(n = 20)

timecourses \%>\%
 tidyr::unnest(measurements) \%>\%
 tidyr::nest(-model, .key = "measurements")

sigmoids <- fit_timecourses_tensorflow(timecourses, model = "sigmoid", n_initializations = 50, use_prior = FALSE)
impulses <- fit_timecourses_tensorflow(timecourse_subset, model = "impulse", n_initializations = 50, use_prior = FALSE)

sigmoids <- fit_timecourses_tensorflow(timecourse_subset, model = "sigmoid", n_initializations = 60, use_prior = TRUE, verbose = TRUE)
impulses <- fit_timecourses_tensorflow(timecourse_subset, model = "impulse", n_initializations = 60, use_prior = TRUE, verbose = TRUE)

}
