% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impulse_fitting.R
\name{estimate_timecourse_params_tf}
\alias{estimate_timecourse_params_tf}
\title{Fit Timecourses}
\usage{
estimate_timecourse_params_tf(
  measurements,
  model = "sigmoid",
  n_initializations = 100,
  max_n_initializations = 1000,
  use_prior = TRUE,
  prior_pars = c(v_sd = 1.2, rate_shape = 2, rate_scale = 0.25, time_shape = 2,
    time_scale = 15),
  fit_intercept = FALSE,
  learning_rate = 0.1,
  verbose = FALSE
)
}
\arguments{
\item{measurements}{a tibble containing:
\describe{
  \item{tc_id}{a unique indicator for each timecourse"}
  \item{time}{a numeric predictor variable}
  \item{abundance}{a numeric response variable}
  \item{noise (optional)}{if provided, a numeric variable specifying the
  noise estimate for each observation. This will be used to calculate the
  Gaussian likelihood if priors are used, or to calculate a weighted
  sum-of-squares estimate if priors are not used}
  }}

\item{model}{model to fit:
\describe{
  \item{sigmoid}{one sigmoidal response}
  \item{impulse}{two sigmoidal responses}
  }}

\item{n_initializations}{Number of initializations to use for each timecourse.}

\item{max_n_initializations}{Maximum number of initializations that can be used}

\item{use_prior}{If FALSE, fit least squares. If TRUE, add priors for a MAP estimate.}

\item{prior_pars}{Named numeric vector of parameters to use for priors (if use_prior is TRUE)
\describe{
  \item{v_sd}{Gaussian SD of assymptotes (v_inter and v_final)}
  \item{rate_shape}{Shape of rate Gamma}
  \item{rate_scale}{Scale of rate Gamma}
  \item{time_shape}{Shape of t_rise and t_fall - t_rise Gamma}
  \item{time_scale}{Scale of t_rise and t_fall - t_rise Gamma}
  }}

\item{fit_intercept}{If TRUE, the intercept will be fit, if FALSE, the intercept will be constrainted to zero}

\item{learning_rate}{learning rate for the Adams optimizer}

\item{verbose}{if TRUE then print additional information}
}
\value{
a timecourse list:
\describe{
  \item{invalid_timecourse_fits}{tibble of parameter initializations for initializations
  that went to NaN for debugging}
  \item{loss}{tibble of losses for each tc_id and inititalization (init_id)}
  \item{parameters}{tibble of final parameters for each tc_id and initialization (init_id)}
  }
}
\description{
Fit Timecourses
}
\examples{

library(dplyr)
library(impulse)
auto_config_tf()

set.seed(123)
timecourses <- simulate_timecourses(n = 2, observation_level_noise = 0.5)

timecourses \%>\%
  tidyr::unnest_legacy(measurements) \%>\%
  # separate by true model
  tidyr::nest_legacy(-true_model, .key = "measurements") \%>\%
  # fit all models to each timecourse
  tidyr::crossing(tibble::tibble(model = c("sigmoid", "impulse"))) \%>\%
  dplyr::mutate(timecourse_params = purrr::map2(
    measurements,
    model,
    estimate_timecourse_params_tf,
    n_initializations = 20,
    use_prior = TRUE,
    fit_intercept = TRUE
    ))

}
