% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_comparison.R
\name{impulse_sigmoid_comparison}
\alias{impulse_sigmoid_comparison}
\title{Impulse Sigmoid Comparison}
\usage{
impulse_sigmoid_comparison(timecourse_losses, fdr_cutoff = 0.001)
}
\arguments{
\item{timecourse_losses}{table containing tc_id, model and logLik}

\item{fdr_cutoff}{cutoff for accepting an impulse over a sigmoid fit}
}
\description{
Impulse Sigmoid Comparison
}
