% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{validate_priors}
\alias{validate_priors}
\title{Validate Parameters}
\usage{
validate_priors(model, prior_pars)

validate_priors(model, prior_pars)
}
\arguments{
\item{model}{name of model to simulate}

\item{prior_pars}{prior parameters for model}

\item{model}{name of model to simulate}

\item{prior_pars}{prior parameters for model}
}
\description{
Check that all required prior parameters are provided and have valid values

Check that all required prior parameters are provided and have valid values
}
