% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{validate_priors}
\alias{validate_priors}
\title{Validate Parameters}
\usage{
validate_priors(model, prior_pars)
}
\arguments{
\item{model}{model to fit:
\describe{
  \item{sigmoid}{one sigmoidal response}
  \item{impulse}{two sigmoidal responses}
  }}

\item{prior_pars}{Named numeric vector of parameters to use for priors (if use_prior is TRUE)
\describe{
  \item{v_sd}{Gaussian SD of assymptotes (v_inter and v_final)}
  \item{rate_shape}{Shape of rate Gamma}
  \item{rate_scale}{Scale of rate Gamma}
  \item{time_shape}{Shape of t_rise and t_fall - t_rise Gamma}
  \item{time_scale}{Scale of t_rise and t_fall - t_rise Gamma}
  }}
}
\description{
Check that all required prior parameters are provided and have valid values
}
